refactor(finance): move GracePeriod type to common
